refactor(react-icons-file-type): split FileTypeIcon into v9 state/style/render hooks
Extract prop construction into a shared useFileTypeIcon state hook plus a
renderFileTypeIcon render hook (headless/useFileTypeIcon). Both the headless
and Griffel-styled FileTypeIcon now compose these hooks; the styled variant
inserts a useFileTypeIconStyles style hook that merges the Griffel class onto
the resolved state, instead of wrapping the headless component. Aligns the
package with the Fluent v9 state/style/render pattern. No public API or
behavior change.
