+++ packages/react-icons-file-type/src/FileTypeIcon.styles.ts
@@ -1,12 +1,13 @@
-import { makeStyles } from '@griffel/react';
+import { makeStyles, mergeClasses } from '@griffel/react';
+import type { FileTypeIconState } from './headless/useFileTypeIcon';
 
 /**
  * Static styles for the `FileTypeIcon` image. Reproduces the box behavior the legacy
  * v8 icon registry applied (`overflow: hidden`) plus sensible image rendering defaults.
  * The dynamic `width`/`height` (which depend on the `size` prop) are applied inline by
- * the component.
+ * the state hook.
  */
-export const useFileTypeIconStyles = makeStyles({
+const useStyles = makeStyles({
   root: {
     display: 'inline-block',
     overflowX: 'hidden',
@@ -14,3 +15,14 @@ export const useFileTypeIconStyles = makeStyles({
     objectFit: 'contain',
   },
 });
+
+/**
+ * Style hook: layers the Griffel `root` class onto a resolved {@link FileTypeIconState},
+ * merging ahead of any consumer-provided `className`. Mutates and returns the same state,
+ * following the Fluent v9 style-hook convention.
+ */
+export function useFileTypeIconStyles(state: FileTypeIconState): FileTypeIconState {
+  const styles = useStyles();
+  state.className = mergeClasses(styles.root, state.className);
+  return state;
+}
